Until now, dynamic parameters where implemented either abusing pfields, or creating a table per event scheduled and using that table to communicate with the event. Pfields can still be used but this seems a hacky method since in the language definition pfields are treated as init (constant) values. They need to be assigned to a kvariable to work and the method relies on the pwrite opcode, which uses what seems to be a loophole in the csound api and could change in a future revision. With this change all dynamic parameters are implemented using a big table. This big table is allocated at the start of a session (its the same online and offline) and whenever an instrument defines dynamic controls it is given a slice (of fixed size) of that table. Dynamic controls are assigned an index within that slice and read from that big table. This method is very efficient since there are no allocations and the table can even be accessed from python without calling csound.
When rendering offline the same mechanism is used but instead of being
given a slice directly, a token is assigned to the event and this
token points then to a slice. This makes it possible to use a pool of
slices which is self regulated within csound without python being
in the middle (this is of course necessary since during offline rendering
there is no python communication).

The .get method of synths has been removed. This relied on the same
experimental api usage as setp and has been removed in favor of more
future-proof methods like buses, channels or OSC

Previously an Instr would generate a fixed body, independent of
under which rendered it was defined. Now the renderer generates the
body for the instr, since it knows best how to interact with the
instr. For real time rendering, for example, the control slot assignment
is done by the host in python. For offline rendering the assignment
mechanism is left to csound itself. This needs different interface
code for each renderer and thus it is left to the rendering stage.

An Engine now tracks changes to the unserlying csound process and can
determine if it needs to sync with it. If started with autosync (now
off by default) then any operation which relie on previous dependencies
(for example scheduling an event of an instrument needs to
wait until the reified version of that instrument has been compiled for
the needed priority) will sync as needed and there is no need to manually
sync with the engine. This makes it more efficient in general since
operations which modify the engine but are not dependent of each other
can be bundled automatically and will not incur into unneeded syncs

* csoundlib: events and score lines can have comments
*