Big rewrite concerning: * named parameters, both for online and offline rendering * aliases * autosync
Named Parameters
----------------

There are two strategies for named controls. One is to use pfields
for both init and dynamic params. Pfields, using the opcode 'pwrite',
can be used to implement dynamic controls in a synth. This is very
efficient both for initialization and automation, less so for set
operations, which need to be done via csound. The other strategy
is to use a big global table, using slots for named params. This is
not so efficient for initialization but it is very efficient for
set and automation. Set operations can be done directly within python,
automation is very similar to dynamic pfields. The method of allocating
a table per instrument has been removed. The strategy can be decided per
instrument, using 'useDynamicPfields' parameter. If True, pfields are
used for all parameters; if False, pfields are used for init time params
and a global table is used for dynamic params

Aliases
-------

Any named parameter can have an alias. This makes it possible to define
names which do not follow the naming convention of csound. It is used,
for example, when calling `playSample`, which has args such as
speed or pan. Within the returned synth it is possible to modify these
parameters by using those same names instead of kspeed or kpan.

Autosync
--------

Any operation which modifies the state of the engine is tracked. Any
operation (sched, for example) which might rely on previous modification
or allocation of resources will issue a sync command before executing. If
the engine needs to sync then the operation is performed. Any later
call to sync will simply return, as long as the engine is not modified
(via compilation, creation of tables, etc)

Other
-----

* interact now generates correct ui for synth / group
* an offline renderer keeps track of all rendered jobs
*
*