Reorganization of the event/synth class structure. A live event (a Synth) is subclass of an offline event (a SchedEvent, for scheduled event). They all can set parameters, automate, stop, etc.
SessionEvent, which was used to express an event TO BE scheduled,
has been renamed to "Event".

Lots of refactoring after checking with mypy. This is mostly about
making the code less dynamic, which should help to understand it
better in the future, and also make refactoring easier. mypy also
drew attention to lots of missing error checking (things that could be
None but we "sort" of know that they should'n, from the context; now
we add checks, which should catch errors in future refactoring).